+++ cookbook/apps/repo_acitivities/api.py
@@ -1,12 +1,19 @@
 import typing
 
-import marvin
 import uvicorn
+from ai import say_you_are_healthy
 from config import settings
 from fastapi import FastAPI, Request
 from gh_util.types import GitHubWebhookRequest
 from pydantic_core import from_json
 from tasks import handle_repo_request
+
+
+class FasterRequest(Request):
+    async def json(self) -> typing.Any:
+        if not hasattr(self, "_json"):
+            self._json = from_json(await self.body())
+        return self._json
 
 
 def save_request(request: GitHubWebhookRequest):
@@ -18,23 +25,6 @@ def save_request(request: GitHubWebhookRequest):
         / f"{request.event.action}_{request.headers.event}_{request.headers.delivery}.json"
     )
     event_log_path.write_text(request.model_dump_json(indent=2, exclude_none=True))
-
-
-class FasterRequest(Request):
-    async def json(self) -> typing.Any:
-        if not hasattr(self, "_json"):
-            self._json = from_json(await self.body())
-        return self._json
-
-
-@marvin.fn(model_kwargs={"model": "gpt-3.5-turbo", "temperature": 1.2})
-def say_you_are_healthy(in_the_style_of: str = "Top Boy (UK tv show)") -> str:
-    """give an extremely short message that you are healthy `in_the_style_of`
-    a given person or character.
-
-    For example, if `in_the_style_of` is "Hagrid (Harry Potter)",
-    > "Bloody hell, I haven't felt this good since Dumbledore gave me a dragon egg!"
-    """
 
 
 app = FastAPI()
